+++ packages/controls/js/libs/code-control/index.js
@@ -392,7 +392,7 @@ const CodeControl = ({
 															.CompletionItemKind
 															.Class,
 														insertText:
-															'.block {\n\t$0\n}',
+															'.block {\n\t$0\n}\n',
 														insertTextRules:
 															monaco.languages
 																.CompletionItemInsertTextRule
@@ -402,10 +402,43 @@ const CodeControl = ({
 															'blockera'
 														),
 														detail: __(
-															'Blockera Block Selector',
+															'Current Block',
 															'blockera'
 														),
 														sortText: '.block',
+														range: {
+															startLineNumber:
+																position.lineNumber,
+															startColumn:
+																position.column -
+																1,
+															endLineNumber:
+																position.lineNumber,
+															endColumn:
+																position.column,
+														},
+													},
+													{
+														label: '.block:hover',
+														kind: monaco.languages
+															.CompletionItemKind
+															.Class,
+														insertText:
+															'.block:hover {\n\t$0\n}\n',
+														insertTextRules:
+															monaco.languages
+																.CompletionItemInsertTextRule
+																.InsertAsSnippet,
+														documentation: __(
+															'Target the current block on hover',
+															'blockera'
+														),
+														detail: __(
+															'Current Block on Hover',
+															'blockera'
+														),
+														sortText:
+															'.block:hover',
 														range: {
 															startLineNumber:
 																position.lineNumber,
@@ -487,6 +520,17 @@ const CodeControl = ({
 
 						if (value !== editor.getValue()) {
 							editor.setValue(value);
+						}
+
+						// Set cursor position between curly braces for CSS
+						if (lang === 'css' && value === '.block {\n    \n}\n') {
+							const position = editor.getPosition();
+							if (position) {
+								editor.setPosition({
+									lineNumber: 2,
+									column: 4,
+								});
+							}
 						}
 					}}
 				/>