+++ packages/editor/js/canvas-editor/helpers.js
@@ -26,9 +26,6 @@ export const getTargets = (version: string): GetTarget => {
 
 	const targets = {
 		header: '.editor-header__center',
-		previewDropdown:
-			'.editor-preview-dropdown, a.components-button[aria-label="View Post"], a.components-button[aria-label="View Page"]',
-		postPreviewElement: '.editor-header__post-preview-button + a',
 	};
 
 	// For WordPress version equals or bigger than 6.6.1 version.
@@ -45,8 +42,5 @@ export const getTargets = (version: string): GetTarget => {
 		header: isLoadedSiteEditor()
 			? '.edit-site-header-edit-mode__center'
 			: '.edit-post-header__center',
-		postPreviewElement:
-			'a[aria-label="View Post"], a[aria-label="View Page"]',
-		previewDropdown: 'div.edit-site-header-edit-mode__preview-options',
 	};
 };