+++ packages/editor/js/canvas-editor/types/index.js
@@ -2,6 +2,4 @@
 
 export type GetTarget = {
 	header: string,
-	previewDropdown: string,
-	postPreviewElement: string,
 };