+++ packages/editor/js/extensions/libs/size/test/aspect-ratio.e2e.cy.js
@@ -45,8 +45,55 @@ describe('Aspect Ratio → Functionality', () => {
 		// Custom
 		cy.getParentContainer('Aspect Ratio').within(() => {
 			cy.get('select').select('custom');
-			cy.get('input').eq(0).type(2);
-			cy.get('input').eq(1).type(5);
+		});
+
+		// Check store
+		getWPDataObject().then((data) => {
+			expect({
+				val: 'custom',
+				width: '1',
+				height: '1',
+			}).to.be.deep.equal(getSelectedBlock(data, 'blockeraRatio'));
+		});
+
+		// Standard 4:3
+		cy.getParentContainer('Aspect Ratio').within(() => {
+			cy.get('select').select('4/3', { force: true });
+		});
+
+		// Check store
+		getWPDataObject().then((data) => {
+			expect({
+				val: '4/3',
+				width: '',
+				height: '',
+			}).to.be.deep.equal(getSelectedBlock(data, 'blockeraRatio'));
+		});
+
+		// Custom
+		cy.getParentContainer('Aspect Ratio').within(() => {
+			cy.get('select').select('custom');
+		});
+
+		// Check store
+		getWPDataObject().then((data) => {
+			expect({
+				val: 'custom',
+				width: '4',
+				height: '3',
+			}).to.be.deep.equal(getSelectedBlock(data, 'blockeraRatio'));
+		});
+
+		// set custom ratio
+		cy.getParentContainer('Aspect Ratio').within(() => {
+			cy.get('input')
+				.eq(0)
+				.clear({ force: true })
+				.type(2, { force: true });
+			cy.get('input')
+				.eq(1)
+				.clear({ force: true })
+				.type(5, { force: true });
 		});
 
 		// Check block